requirements.txt
agate==1.13.0
agate-dbf==0.2.3
agate-excel==0.4.1
agate-sql==0.7.2
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.5.2
babel==2.17.0
backoff==2.2.1
certifi==2025.6.15
cffi==1.17.1
click==8.1.8
cryptography==45.0.4
csvkit==2.1.0
dbfread==2.0.7
distro==1.9.0
et-xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.115.13
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib-metadata==8.5.0
isodate==0.7.2
jiter==0.9.1
leather==0.4.0
networkx==3.1
numpy==1.24.4
olefile==0.47
openai==1.88.0
openpyxl==3.1.5
parsedatetime==2.6
pathtools==0.1.2
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
PyJWT==2.9.0
python-dotenv==1.0.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.2
sniffio==1.3.1
sqlalchemy==2.0.41
starlette==0.44.0
text-unidecode==1.3
tqdm==4.67.1
typing-extensions==4.13.2
uvicorn==0.33.0
xlrd==2.0.2
zipp==3.20.2
